refactor(core): move to_relative_unix_style (+ helpers) to julie-core::paths
Move to_relative_unix_style, strip_unc_prefix, relative_by_normalized_string,
and both strip_normalized_prefix variants into crates/julie-core/src/paths.rs.

strip_unc_prefix is promoted to pub so src/utils/paths.rs can re-import it
privately for relative_within_workspace, which stays in the main crate.
src/utils/paths.rs gains:
  use julie_core::paths::strip_unc_prefix;      // private, for relative_within_workspace
  pub use julie_core::paths::to_relative_unix_style;  // shim for ~16 callers

All existing crate::utils::paths::to_relative_unix_style call sites compile
unchanged — zero call-site edits.

Also removes the orphaned doc block that was accidentally placed before
strip_unc_prefix (it described to_relative_unix_style but was separated from it).

Gates:
  cargo check                   ✓ (zero errors)
  utils::paths::tests           ✓ 11/11 PASS (no assertion changes)
  cargo build                   ✓ (full workspace)
